Refactor: project layout by relocating files to improve overall code organization and maintainability.
