+++ app/GX_integration/data_manager.py
