+++ app/GX_integration/history_manager.py
