+++ app/api/endpoints/histories.py
@@ -16,8 +16,8 @@ import shutil
 
 from app.context import current_api_key
 from app.galaxy import GalaxyClient
-from app.bioblend_server.executor.data_manager import DataManager, CollectionType
-from app.bioblend_server.executor.history_manager import HistoryManager
+from app.GX_integration.data_manager import DataManager, CollectionType
+from app.GX_integration.history_manager import HistoryManager
 from app.api.schemas import dataset, history
 
 from app.exceptions import InternalServerErrorException, BadRequestException
