+++ app/api/endpoints/tools.py
@@ -11,7 +11,7 @@ from app.context import current_api_key
 from app.api.schemas.tool import ToolExecutionResponse
 from app.api.schemas.workflow import OutputDataset, CollectionOutputDataset
 from app.galaxy import GalaxyClient 
-from app.bioblend_server.executor.tool_manager import ToolManager
+from app.GX_integration.tool_manager import ToolManager
 from app.api.socket_manager import ws_manager
 
 from app.exceptions import InternalServerErrorException
