+++ app/orchestration/invocation_tasks.py
@@ -13,7 +13,7 @@ from starlette.concurrency import run_in_threadpool
 from app.orchestration.invocation_cache import InvocationCache
 from app.orchestration.utils import NumericLimits
 from app.galaxy import GalaxyClient
-from app.bioblend_server.executor.workflow_manager import WorkflowManager
+from app.GX_integration.workflows.workflow_manager import WorkflowManager
 
 load_dotenv()
